fix: Restore IO capabilities for hooks and complete crate rename
Restores Embed IO to HookEffects to support FFI host calls required by stop hooks. Completes the rename of exomonad-sidecar to exomonad in the Rust workspace.
