+++ .exomonad/.gitignore
@@ -6,3 +6,5 @@ wasm/
 worktrees/
 sockets/
 logs/
+result
+result-*
