+++ rust/Cargo.toml
@@ -4,8 +4,8 @@ members = [
     "exomonad-shared",
     "exomonad-ui-protocol",
     "exomonad-services",
+    "exomonad",
     "exomonad-runtime",
-    "exomonad-sidecar",
     "exomonad-repl",
     "effector",
     "zellij-gen",
