--- rust/exomonad-sidecar/src/config.rs
@@ -1,356 +1,0 @@
-//! Configuration discovery from .exomonad/config.toml
-
-use anyhow::{Context, Result};
-use exomonad_shared::{PathError, Role, WasmPath};
-use serde::Deserialize;
-use std::path::PathBuf;
-use thiserror::Error;
-use tracing::info;
-
-/// Configuration validation errors.
-#[derive(Debug, Error)]
-pub enum ConfigError {
-    #[error("project directory does not exist: {path}")]
-    ProjectDirNotFound { path: PathBuf },
-
-    #[error("project directory is not a directory: {path}")]
-    ProjectDirNotDirectory { path: PathBuf },
-
-    #[error("HOME environment variable not set")]
-    HomeNotSet,
-
-    #[error("WASM plugin validation failed: {0}")]
-    WasmValidation(#[from] PathError),
-
-    #[error("I/O error: {0}")]
-    Io(#[from] std::io::Error),
-}
-
-/// Sidecar configuration.
-#[derive(Debug, Clone, Deserialize)]
-pub struct Config {
-    /// Project directory for git operations.
-    #[serde(default = "default_project_dir")]
-    pub project_dir: PathBuf,
-
-    /// Agent role (dev, tl, pm).
-    #[serde(default)]
-    pub role: Role,
-}
-
-impl Default for Config {
-    fn default() -> Self {
-        Self {
-            project_dir: default_project_dir(),
-            role: Role::default(),
-        }
-    }
-}
-
-fn default_project_dir() -> PathBuf {
-    PathBuf::from(".")
-}
-
-/// Validated configuration wrapper.
-///
-/// This type guarantees that all configuration fields have been validated:
-/// - project_dir exists and is a directory
-/// - WASM path exists, is a file, and has .wasm extension (stored as WasmPath)
-#[derive(Debug, Clone)]
-pub struct ValidatedConfig {
-    config: Config,
-    wasm_path: WasmPath,
-}
-
-impl Config {
-    /// Discover configuration from .exomonad/config.toml in current directory.
-    pub fn discover() -> Result<Self> {
-        let path = PathBuf::from(".exomonad/config.toml");
-        Self::from_path(&path)
-    }
-
-    /// Load configuration from a specific path.
-    pub fn from_path(path: &PathBuf) -> Result<Self> {
-        info!(path = %path.display(), "Loading config");
-
-        let content = std::fs::read_to_string(path)
-            .with_context(|| format!("Failed to read config file: {}", path.display()))?;
-
-        let config: Config = toml::from_str(&content)
-            .with_context(|| format!("Failed to parse config file: {}", path.display()))?;
-
-        info!(project_dir = %config.project_dir.display(), "Config loaded");
-
-        Ok(config)
-    }
-
-    /// Create a default config (for when no config file exists).
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Validate configuration and return a ValidatedConfig.
-    ///
-    /// This method performs all validation checks and consumes self to return
-    /// a ValidatedConfig that guarantees all fields are valid.
-    pub fn validate(self) -> Result<ValidatedConfig, ConfigError> {
-        // Check project_dir exists
-        if !self.project_dir.exists() {
-            return Err(ConfigError::ProjectDirNotFound {
-                path: self.project_dir.clone(),
-            });
-        }
-
-        // Check project_dir is a directory
-        if !self.project_dir.is_dir() {
-            return Err(ConfigError::ProjectDirNotDirectory {
-                path: self.project_dir.clone(),
-            });
-        }
-
-        // Check if it's a git repo (has .git directory)
-        // This is a warning, not a hard error - sidecar can still run,
-        // but git operations won't work
-        let git_dir = self.project_dir.join(".git");
-        if !git_dir.exists() {
-            info!(
-                path = %self.project_dir.display(),
-                "Warning: project directory is not a git repository"
-            );
-        }
-
-        // Resolve and validate WASM path using WasmPath newtype
-        let wasm_path_buf = self.wasm_path_internal()?;
-        let wasm_path = WasmPath::try_from(wasm_path_buf)?;
-
-        Ok(ValidatedConfig {
-            config: self,
-            wasm_path,
-        })
-    }
-
-    /// Internal helper for WASM path resolution (returns ConfigError).
-    fn wasm_path_internal(&self) -> Result<PathBuf, ConfigError> {
-        let home = std::env::var("HOME").map_err(|_| ConfigError::HomeNotSet)?;
-        Ok(PathBuf::from(home)
-            .join(".exomonad/wasm")
-            .join(format!("wasm-guest-{}.wasm", self.role)))
-    }
-}
-
-impl ValidatedConfig {
-    /// Get the validated project directory.
-    pub fn project_dir(&self) -> &PathBuf {
-        &self.config.project_dir
-    }
-
-    /// Get the validated role.
-    pub fn role(&self) -> Role {
-        self.config.role
-    }
-
-    /// Get the WASM path as a PathBuf.
-    pub fn wasm_path_buf(&self) -> PathBuf {
-        self.wasm_path.as_path().to_path_buf()
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::io::Write;
-    use tempfile::tempdir;
-
-    #[test]
-    fn test_parse_minimal_config() {
-        // Empty config should use defaults
-        let toml = "";
-        let config: Config = toml::from_str(toml).unwrap();
-        assert_eq!(config.project_dir, PathBuf::from("."));
-        assert_eq!(config.role, Role::Dev);
-    }
-
-    #[test]
-    fn test_parse_full_config() {
-        let toml = r#"
-project_dir = "/home/user/project"
-"#;
-        let config: Config = toml::from_str(toml).unwrap();
-        assert_eq!(config.project_dir, PathBuf::from("/home/user/project"));
-    }
-
-    #[test]
-    fn test_from_path() {
-        let dir = tempdir().unwrap();
-        let config_path = dir.path().join("config.toml");
-
-        let mut file = std::fs::File::create(&config_path).unwrap();
-        writeln!(file, r#"project_dir = "/tmp/test""#).unwrap();
-
-        let config = Config::from_path(&config_path).unwrap();
-        assert_eq!(config.project_dir, PathBuf::from("/tmp/test"));
-    }
-
-    #[test]
-    fn test_default() {
-        let config = Config::default();
-        assert_eq!(config.project_dir, PathBuf::from("."));
-        assert_eq!(config.role, Role::Dev);
-    }
-
-    #[test]
-    fn test_wasm_path() {
-        let config = Config {
-            role: Role::TL,
-            project_dir: PathBuf::from("."),
-        };
-
-        let expected =
-            PathBuf::from(std::env::var("HOME").unwrap()).join(".exomonad/wasm/wasm-guest-tl.wasm");
-
-        assert_eq!(config.wasm_path_internal().unwrap(), expected);
-    }
-
-    #[test]
-    fn test_wasm_path_dev_role() {
-        let config = Config::default();
-
-        let expected = PathBuf::from(std::env::var("HOME").unwrap())
-            .join(".exomonad/wasm/wasm-guest-dev.wasm");
-
-        assert_eq!(config.wasm_path_internal().unwrap(), expected);
-    }
-
-    #[test]
-    fn test_validate_project_dir_not_found() {
-        let config = Config {
-            project_dir: PathBuf::from("/nonexistent/path/to/project"),
-            role: Role::Dev,
-        };
-
-        let result = config.validate();
-        assert!(matches!(
-            result,
-            Err(ConfigError::ProjectDirNotFound { .. })
-        ));
-    }
-
-    #[test]
-    fn test_validate_project_dir_not_directory() {
-        // Create a temporary file (not a directory)
-        let dir = tempdir().unwrap();
-        let file_path = dir.path().join("notadir");
-        std::fs::File::create(&file_path).unwrap();
-
-        let config = Config {
-            project_dir: file_path,
-            role: Role::Dev,
-        };
-
-        let result = config.validate();
-        assert!(matches!(
-            result,
-            Err(ConfigError::ProjectDirNotDirectory { .. })
-        ));
-    }
-
-    #[test]
-    fn test_validate_not_git_repo() {
-        // Create a directory without .git
-        let dir = tempdir().unwrap();
-
-        let config = Config {
-            project_dir: dir.path().to_path_buf(),
-            role: Role::Dev,
-        };
-
-        // Not being a git repo is now a warning, not an error
-        // Validation may pass if WASM exists, or fail if it doesn't
-        let result = config.validate();
-        match result {
-            Ok(_) => {
-                // WASM exists in dev environment
-            }
-            Err(ConfigError::WasmValidation(_)) => {
-                // Expected if WASM not installed (PathError::NotFound)
-            }
-            Err(e) => {
-                panic!("Expected WasmValidation or success, got: {:?}", e);
-            }
-        }
-    }
-
-    #[test]
-    fn test_validate_wasm_not_found() {
-        // Create a directory with .git
-        let dir = tempdir().unwrap();
-        let git_dir = dir.path().join(".git");
-        std::fs::create_dir(&git_dir).unwrap();
-
-        let config = Config {
-            project_dir: dir.path().to_path_buf(),
-            role: Role::Dev,
-        };
-
-        // WASM might exist at ~/.exomonad/wasm/wasm-guest-dev.wasm in dev environment
-        let result = config.validate();
-        // Either WasmNotFound (doesn't exist) or success (exists in dev env)
-        match result {
-            Ok(_) => {
-                // WASM exists in dev environment, test passes
-            }
-            Err(ConfigError::WasmValidation(_)) => {
-                // Expected if WASM not installed
-            }
-            Err(e) => {
-                panic!("Expected WasmNotFound or success, got: {:?}", e);
-            }
-        }
-    }
-
-    #[test]
-    fn test_validate_success() {
-        // Create a directory with .git
-        let dir = tempdir().unwrap();
-        let git_dir = dir.path().join(".git");
-        std::fs::create_dir(&git_dir).unwrap();
-
-        // Create WASM file
-        let home = std::env::var("HOME").unwrap();
-        let wasm_dir = PathBuf::from(&home).join(".exomonad").join("wasm");
-        std::fs::create_dir_all(&wasm_dir).unwrap();
-        let wasm_path = wasm_dir.join("wasm-guest-test-validate.wasm");
-        std::fs::write(&wasm_path, b"fake wasm").unwrap();
-
-        let _config = Config {
-            project_dir: dir.path().to_path_buf(),
-            role: Role::Dev,
-        };
-
-        // Skip this test - it requires actual WASM files to exist
-        // The test_validate_wasm_not_found test already covers the failure case
-    }
-
-    #[test]
-    fn test_validated_config_methods() {
-        // This test requires a valid config, so we'll just test the methods exist
-        // and have the right signatures
-        let dir = tempdir().unwrap();
-        let git_dir = dir.path().join(".git");
-        std::fs::create_dir(&git_dir).unwrap();
-
-        let config = Config {
-            project_dir: dir.path().to_path_buf(),
-            role: Role::Dev,
-        };
-
-        // If validation passes, test the methods
-        // (This will fail if WASM doesn't exist, which is expected in test env)
-        if let Ok(validated) = config.validate() {
-            assert_eq!(validated.role(), Role::Dev);
-            assert_eq!(validated.project_dir(), dir.path());
-            let wasm = validated.wasm_path_buf();
-            assert!(wasm.to_str().unwrap().ends_with("wasm-guest-dev.wasm"));
-        }
-    }
-}
